# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UNILIV Admin API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: getProperties
      tags: [properties]
      summary: List properties
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
      responses:
        "200":
          description: Properties list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertiesListResponse"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyBody"
      responses:
        "201":
          description: Property created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyResponse"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get property by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyResponse"
    put:
      operationId: updateProperty
      tags: [properties]
      summary: Update property
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyBody"
      responses:
        "200":
          description: Property updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyResponse"
    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete property
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── ROOMS ────────────────────────────────────────────
...
  /residents:
    get:
      operationId: getResidents
      tags: [residents]
      summary: List residents
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
        - $ref: "#/components/parameters/PropertyIdParam"
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Residents list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidentsListResponse"
    post:
      operationId: createResident
      tags: [residents]
      summary: Create resident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResidentBody"
      responses:
        "201":
          description: Resident created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidentResponse"
  /residents/{id}:
    get:
      operationId: getResident
      tags: [residents]
      summary: Get resident by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Resident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidentResponse"
    put:
      operationId: updateResident
      tags: [residents]
      summary: Update resident
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResidentBody"
      responses:
        "200":
          description: Resident updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidentResponse"
    delete:
      operationId: deleteResident
      tags: [residents]
      summary: Delete resident
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /residents/{id}/ledger:
    get:
      operationId: getResidentLedger
      tags: [residents]
      summary: Get resident ledger entries
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Ledger entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerListResponse"
    post:
      operationId: createLedgerEntry
      tags: [residents]
      summary: Create ledger entry
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLedgerEntryBody"
      responses:
        "201":
          description: Ledger entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerEntryResponse"
  /residents/{id}/payments:
    get:
      operationId: getResidentPayments
      tags: [residents]
      summary: Get resident payments
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Payments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentsListResponse"
    post:
      operationId: createPayment
      tags: [residents]
      summary: Create payment
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"

  # ─── COMPLAINTS ───────────────────────────────────────
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
    SearchParam:
      name: search
      in: query
      schema:
        type: string
    PropertyIdParam:
      name: propertyId
      in: query
      schema:
        type: string
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [total, page, limit, totalPages]

    # ─── AUTH ─────────────────────────────────────────────
...
    PropertyDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        totalBeds:
          type: integer
        status:
          type: string
        wardenId:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        amenities:
          type: array
          items:
            type: string
        occupiedBeds:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, address, city, state, pincode, totalBeds, status, amenities, occupiedBeds, createdAt, updatedAt]
    PropertyResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/PropertyDto"
      required: [success, data]
    PropertiesListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/PropertyDto"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [success, data, meta]
    CreatePropertyBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        lat:
          type: number
        lng:
          type: number
        totalBeds:
          type: integer
        status:
          type: string
        wardenId:
          type: string
        phone:
          type: string
        email:
          type: string
        amenities:
          type: array
          items:
            type: string
      required: [name, address, city, state, pincode, totalBeds]

    # ─── ROOM ─────────────────────────────────────────────
...
    ResidentDto:
      type: object
      properties:
        id:
          type: string
        propertyId:
          type: string
        roomId:
          type: string
          nullable: true
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dob:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        photo:
          type: string
          nullable: true
        college:
          type: string
          nullable: true
        course:
          type: string
          nullable: true
        parentName:
          type: string
          nullable: true
        parentPhone:
          type: string
          nullable: true
        dietaryPref:
          type: array
          items:
            type: string
        checkInDate:
          type: string
          nullable: true
        checkOutDate:
          type: string
          nullable: true
        planType:
          type: string
          nullable: true
        monthlyRent:
          type: number
          nullable: true
        securityDeposit:
          type: number
          nullable: true
        status:
          type: string
        createdAt:
          type: string
        propertyName:
          type: string
          nullable: true
        roomNumber:
          type: string
          nullable: true
      required: [id, propertyId, name, email, phone, dietaryPref, status, createdAt]
    ResidentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/ResidentDto"
      required: [success, data]
    ResidentsListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/ResidentDto"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [success, data, meta]
    CreateResidentBody:
      type: object
      properties:
        propertyId:
          type: string
        roomId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dob:
          type: string
        gender:
          type: string
        college:
          type: string
        course:
          type: string
        parentName:
          type: string
        parentPhone:
          type: string
        parentEmail:
          type: string
        dietaryPref:
          type: array
          items:
            type: string
        checkInDate:
          type: string
        checkOutDate:
          type: string
        planType:
          type: string
        monthlyRent:
          type: number
        securityDeposit:
          type: number
        status:
          type: string
      required: [propertyId, name, email, phone]

    # ─── LEDGER ───────────────────────────────────────────
    LedgerEntryDto:
      type: object
      properties:
        id:
          type: string
        residentId:
          type: string
        type:
          type: string
        amount:
          type: number
        description:
          type: string
        dueDate:
          type: string
          nullable: true
        isPaid:
          type: boolean
        paidOn:
          type: string
          nullable: true
        reference:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, residentId, type, amount, description, isPaid, createdAt]
    LedgerEntryResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/LedgerEntryDto"
      required: [success, data]
    LedgerListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/LedgerEntryDto"
      required: [success, data]
    CreateLedgerEntryBody:
      type: object
      properties:
        type:
          type: string
        amount:
          type: number
        description:
          type: string
        dueDate:
          type: string
        isPaid:
          type: boolean
        reference:
          type: string
      required: [type, amount, description]

    # ─── PAYMENT ──────────────────────────────────────────
    PaymentDto:
      type: object
      properties:
        id:
          type: string
        residentId:
          type: string
        amount:
          type: number
        mode:
          type: string
        status:
          type: string
        reference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, residentId, amount, mode, status, createdAt]
    PaymentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/PaymentDto"
      required: [success, data]
    PaymentsListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/PaymentDto"
      required: [success, data]
    CreatePaymentBody:
      type: object
      properties:
        amount:
          type: number
        mode:
          type: string
        status:
          type: string
        reference:
          type: string
        notes:
          type: string
      required: [amount, mode]

    # ─── COMPLAINT ────────────────────────────────────────
...