# lib/api-zod/src/generated/types/paymentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UNILIV Admin API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentDto } from "./paymentDto";

export interface PaymentResponse {
  success: boolean;
  data: PaymentDto;
}